update: services and handlers logic improvement

+++ reach-goals-app/api/tag/[tagID].js
@@ -5,55 +5,62 @@ import {
 } from '../../server/services/tag.service.js'
 import { formatObject } from '../utils/utils.js'
 
+const ALLOWED_METHODS = ['GET', 'PUT', 'DELETE']
+
 const handler = async (req, res) => {
-    if (req.method === 'GET') {
-        const { tagID } = req.query
-        const tag = await getTag(tagID)
+    const { tagID } = req.query
 
-        if (tag) {
+    if (!ALLOWED_METHODS.includes(req.method)) {
+        return res.status(405).json({
+            error: 'Method not allowed. Check the type of method sended',
+        })
+    }
+
+    try {
+        if (req.method === 'GET') {
+            const tag = await getTag(tagID)
+
             return res.status(200).json(Array.isArray(tag) ? tag : [tag])
-        } else {
-            return res.status(500).json({ error: 'Failed to get a tag' })
         }
+
+        if (req.method === 'PUT') {
+            const { name, color } = req.body
+
+            if (!name || !color) {
+                return res
+                    .status(400)
+                    .json({ error: 'Name/Color is required.' })
+            }
+
+            const rawObject = { name, color }
+
+            const formattedData = formatObject(rawObject)
+            const tag = await updateTag(tagID, formattedData)
+
+            return res.status(201).json(tag)
+        }
+
+        if (req.method === 'DELETE') {
+            await deleteTag(tagID)
+            return res.status(200).json({ message: 'Tag deleted successfully' })
+        }
+    } catch (error) {
+        const responseStatus = res.status
+
+        if (responseStatus === '401') {
+            return res.status(401).json({
+                restartDemo: true,
+                message: 'Unauthorized. Demo session expired. Try a new login',
+                error,
+            })
+        }
+
+        return res.status(500).json({
+            error:
+                `Failed to process request - ${action}: ${error.message}` ||
+                'Internal Server Error',
+        })
     }
-
-    if (req.method === 'PUT') {
-        const { tagID } = req.query
-        const { name, color } = req.body
-
-        if (!name || !color) {
-            return res.status(400).json({ error: 'Name/Color is required.' })
-        }
-
-        const rawObject = { name, color }
-
-        const formattedData = formatObject(rawObject)
-        const tag = await updateTag(tagID, formattedData)
-
-        try {
-            if (tag) return res.status(201).json(tag)
-        } catch (err) {
-            console.error('Error update tag:', err)
-            return res
-                .status(500)
-                .json({ error: err.message || 'Failed updating tags' })
-        }
-    }
-
-    if (req.method === 'DELETE') {
-        const { tagID } = req.query
-        const tag = await deleteTag(tagID)
-
-        if (tag) {
-            return res.status(200).json(tag)
-        } else {
-            return res.status(500).json({ error: 'Failed to delete this tag' })
-        }
-    }
-
-    return res
-        .status(405)
-        .json({ error: 'Method not allowed. Check the type of method sended' })
 }
 
 export default handler
